hmm perf issues again... this sidebar is rough
