+++ packages/frontend/app/components/app/sidebar/chats/data/index.hbs
@@ -1,0 +1,7 @@
+{{yield
+  (hash
+    chats=this.chats
+    numberOffline=this.numberOffline
+    showOfflineCounter=this.showOfflineCounter
+  )
+}}
