+++ packages/frontend/app/components/app/sidebar/chats/list.hbs
@@ -1,0 +1,28 @@
+{{#each @chats as |chat|}}
+{{!--
+{{#vertical-collection
+  @chats
+  estimateHeight=40 staticHeight=true bufferSize=5
+  as |chat|
+}}
+--}}
+  {{#if (eq chat 'add-contact')}}
+    <div data-test-sidebar-contacts-header class='menu-label'>
+      {{t 'ui.sidebar.contacts.title'}}
+
+      <a
+        data-test-add-friend
+        class='button button-xs'
+        href={{url-for 'add-friend'}}
+        {{on 'click' (handle-sidebar-click (transition-to 'add-friend'))}}
+      >
+        <span class='icon'>{{fa-icon 'plus'}}</span>
+      </a>
+    </div>
+  {{else if (or (is-contact chat) (is-current-user chat))}}
+    <App::Sidebar::Chats::Contact @contact={{chat}} />
+  {{else if (is-channel chat)}}
+    <App::Sidebar::Chats::Channel @channel={{chat}} />
+  {{/if}}
+{{!-- {{/vertical-collection}} --}}
+{{/each}}
