+++ packages/frontend/app/components/app/sidebar/chats/offline-counter.hbs
@@ -1,0 +1,3 @@
+<em data-test-offline-count class='offline'>
+  <small>{{t 'ui.sidebar.contacts.numOffline' num=@numberOffline}}</small>
+</em>
