+++ packages/frontend/app/components/app/sidebar/index.hbs
@@ -20,43 +20,6 @@
   </div>
 
   {{#if this.isLoggedIn}}
-    <footer>
-      <hr>
-      <div>
-        {{this.name}}
-
-        <nav>
-          <ModalStatic @name='search' as |isActive actions|>
-            <button type='button' {{on 'click' actions.toggle}}>
-              {{fa-icon 'search' prefix='fas'}}
-            </button>
-          </ModalStatic>
-
-          <a
-            href={{url-for 'contacts'}}
-            title={{t 'routes.contacts'}}
-            {{on 'click' (handle-sidebar-click (transition-to 'contacts'))}}
-          >
-            {{fa-icon 'address-book' prefix='fas'}}
-          </a>
-
-          <a
-            href={{url-for 'settings'}}
-            title={{t 'routes.settings'}}
-            {{on 'click' (handle-sidebar-click (transition-to 'settings'))}}
-          >
-            {{fa-icon 'sliders-h'}}
-          </a>
-
-          <a
-            href={{url-for 'logout'}}
-            title={{t 'routes.logout'}}
-            {{on 'click' (handle-sidebar-click (transition-to 'logout'))}}
-          >
-            {{fa-icon 'sign-out-alt' prefix='fas'}}
-          </a>
-        </nav>
-      </div>
-    </footer>
+    <App::Sidebar::Footer />
   {{/if}}
 </aside>
